requirements.txt
python-telegram-bot==21.4
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
certifi==2024.8.30
